These commits adress issues #271 #270 and #269
