+++ proteobench/modules/dda_quant_base/module.py
@@ -7,7 +7,6 @@ import os
 from collections import ChainMap
 from dataclasses import asdict
 from tempfile import TemporaryDirectory
-from typing import Tuple
 
 import numpy as np
 import pandas as pd
@@ -21,21 +20,19 @@ from proteobench.io.params.fragger import extract_params as extract_params_fragg
 from proteobench.io.params.maxquant import extract_params as extract_params_maxquant
 from proteobench.io.params.proline import extract_params as extract_params_proline
 from proteobench.io.params.sage import extract_params as extract_params_sage
+from proteobench.score.quant.quantscores import QuantScores
 from proteobench.utils.quant_datapoint import Datapoint
+
 from proteobench.io.parsing.parse_ion import (
-    ParseInputs,
-    aggregate_modification_column,
     load_input_file,
 )
 from proteobench.io.parsing.parse_settings_ion import (
     PRECURSOR_NAME,
-    ParseSettings,
     ParseSettingsBuilder,
 )
-from proteobench.modules.interfaces import ModuleInterface
 
 
-class Module(ModuleInterface):
+class Module:
     """Object is used as a main interface with the Proteobench library within the module."""
 
     def __init__(self):
@@ -53,193 +50,6 @@ class Module(ModuleInterface):
         "Sage": extract_params_sage,
         "FragPipe": extract_params_fragger,
     }
-
-    @staticmethod
-    def convert_replicate_to_raw(replicate_to_raw: dict) -> pd.DataFrame:
-        """Converts replicate_to_raw dictionary into a dataframe."""
-        replicate_to_raw_df = pd.DataFrame(replicate_to_raw.items(), columns=["Group", "Raw file"])
-        replicate_to_raw_df = replicate_to_raw_df.explode("Raw file")
-        return replicate_to_raw_df
-
-    def generate_intermediate(
-        self,
-        filtered_df: pd.DataFrame,
-        replicate_to_raw: dict,
-        parse_settings: ParseSettings,
-    ) -> pd.DataFrame:
-        # select columns which are relavant for the statistics
-        # TODO, this should be handled different, probably in the parse settings
-        relevant_columns_df = filtered_df[["Raw file", self.precursor_name, "Intensity"]].copy()
-        replicate_to_raw_df = Module.convert_replicate_to_raw(replicate_to_raw)
-
-        # add column "Group" to filtered_df_p1 using inner join on "Raw file"
-        relevant_columns_df = pd.merge(relevant_columns_df, replicate_to_raw_df, on="Raw file", how="inner")
-
-        quant_df = Module.compute_group_stats(
-            relevant_columns_df,
-            min_intensity=0,
-            precursor=self.precursor_name,
-        )
-
-        species_prec_ion = list(parse_settings.species_dict().values())
-        species_prec_ion.append(self.precursor_name)
-        prec_ion_to_species = filtered_df[species_prec_ion].drop_duplicates()
-        # merge dataframes quant_df and species_quant_df and prec_ion_to_species using pepdidoform as index
-        quant_df_withspecies = pd.merge(quant_df, prec_ion_to_species, on=self.precursor_name, how="inner")
-        species_expected_ratio = parse_settings.species_expected_ratio()
-        res = Module.compute_epsilon(quant_df_withspecies, species_expected_ratio)
-        return res
-
-    @staticmethod
-    def compute_group_stats(
-        relevant_columns_df: pd.DataFrame,
-        min_intensity=0,
-        precursor="precursor ion",
-    ) -> pd.DataFrame:
-        """Method used to precursor statistics, such as number of observations, CV, mean per group etc."""
-
-        # fiter for min_intensity
-        relevant_columns_df = relevant_columns_df[relevant_columns_df["Intensity"] > min_intensity]
-
-        # TODO: check if this is still needed
-        # sum intensity values of the same precursor and "Raw file" using the sum
-        quant_raw_df_int = (
-            relevant_columns_df.groupby([precursor, "Raw file", "Group"])["Intensity"]
-            .agg(Intensity="sum", Count="size")
-            .reset_index()
-        )
-
-        # add column "log_Intensity" to quant_raw_df
-        quant_raw_df_int["log_Intensity"] = np.log2(quant_raw_df_int["Intensity"])
-
-        # compute the mean of the log_Intensity per precursor and "Group"
-        quant_raw_df_count = (quant_raw_df_int.groupby([precursor])).agg(nr_observed=("Raw file", "size"))
-
-        # pivot filtered_df_p1 to wide where index peptide ion, columns Raw file and values Intensity
-
-        intensities_wide = quant_raw_df_int.pivot(index=precursor, columns="Raw file", values="Intensity").reset_index()
-
-        quant_raw_df = (
-            quant_raw_df_int.groupby([precursor, "Group"])
-            .agg(
-                log_Intensity_mean=("log_Intensity", "mean"),
-                log_Intensity_std=("log_Intensity", "std"),
-                Intensity_mean=("Intensity", "mean"),
-                Intensity_std=("Intensity", "std"),
-                Sum=("Intensity", "sum"),
-                nr_obs_group=("Intensity", "size"),
-            )
-            .reset_index()
-        )
-
-        # compute coefficient of variation (CV) of the log_Intensity_mean and log_Intensity_std
-        quant_raw_df["CV"] = quant_raw_df["Intensity_std"] / quant_raw_df["Intensity_mean"]
-        # pivot dataframe wider so for each Group variable there is a column with log_Intensity_mean, log_Intensity_std, Intensity_mean, Intensity_std and CV
-        quant_raw_df = quant_raw_df.pivot(
-            index=precursor,
-            columns="Group",
-            values=[
-                "log_Intensity_mean",
-                "log_Intensity_std",
-                "Intensity_mean",
-                "Intensity_std",
-                "CV",
-            ],
-        ).reset_index()
-
-        quant_raw_df.columns = [f"{x[0]}_{x[1]}" if len(str(x[1])) > 0 else x[0] for x in quant_raw_df.columns]
-
-        quant_raw_df["log2_A_vs_B"] = quant_raw_df["log_Intensity_mean_A"] - quant_raw_df["log_Intensity_mean_B"]
-
-        quant_raw_df = pd.merge(quant_raw_df, intensities_wide, on=precursor, how="inner")
-        quant_raw_df = pd.merge(quant_raw_df, quant_raw_df_count, on=precursor, how="inner")
-        return quant_raw_df
-
-    @staticmethod
-    def compute_epsilon(withspecies, species_expected_ratio):
-        # for all columns named parse_settings.species_dict.values() compute the sum over the rows and add it to a new column "unique"
-        withspecies["unique"] = withspecies[species_expected_ratio.keys()].sum(axis=1)
-
-        # now remove all rows with withspecies["unique"] > 1
-        withspecies = withspecies[withspecies["unique"] == 1]
-
-        # for species in parse_settings.species_dict.values(), set all values in new column "species" to species if withe species is True
-        for species in species_expected_ratio.keys():
-            withspecies.loc[withspecies[species] == True, "species"] = species
-            withspecies.loc[withspecies[species] == True, "log2_expectedRatio"] = np.log2(
-                species_expected_ratio[species]["A_vs_B"]
-            )
-
-        withspecies["epsilon"] = withspecies["log2_A_vs_B"] - withspecies["log2_expectedRatio"]
-        return withspecies
-
-    @staticmethod
-    def get_metrics(df, min_nr_observed=1):
-        # compute mean of epsilon column in df
-        # take abs value of df["epsilon"]
-        # TODO use nr_missing to filter df before computing stats.
-        df_slice = df[df["nr_observed"] >= min_nr_observed]
-        nr_prec = len(df_slice)
-        # median abs unafected by outliers
-        median_abs_epsilon = df_slice["epsilon"].abs().mean()
-        # variance affected by outliers
-        variance_epsilon = df_slice["epsilon"].var()
-        # TODO more concise way to describe distribution of CV's
-        cv_median = (df_slice["CV_A"].median() + df_slice["CV_B"].median()) / 2
-        cv_q75 = (df_slice["CV_A"].quantile(0.75) + df_slice["CV_B"].quantile(0.75)) / 2
-        cv_q90 = (df_slice["CV_A"].quantile(0.9) + df_slice["CV_B"].quantile(0.9)) / 2
-        cv_q95 = (df_slice["CV_A"].quantile(0.95) + df_slice["CV_B"].quantile(0.95)) / 2
-
-        return {
-            min_nr_observed: {
-                "median_abs_epsilon": median_abs_epsilon,
-                "variance_epsilon": variance_epsilon,
-                "nr_prec": nr_prec,
-                "CV_median": cv_median,
-                "CV_q90": cv_q90,
-                "CV_q75": cv_q75,
-                "CV_q95": cv_q95,
-            }
-        }
-
-    def generate_datapoint(
-        self, intermediate: pd.DataFrame, input_format: str, user_input: dict, default_cutoff_min_prec: int = 3
-    ) -> Datapoint:
-        """Method used to compute metadata for the provided result."""
-        current_datetime = datetime.datetime.now()
-        formatted_datetime = current_datetime.strftime("%Y%m%d_%H%M%S_%f")
-
-        if "comments_for_submission" not in user_input.keys():
-            user_input["comments_for_submission"] = ""
-
-        result_datapoint = Datapoint(
-            id=input_format + "_" + user_input["software_version"] + "_" + formatted_datetime,
-            software_name=input_format,
-            software_version=user_input["software_version"],
-            search_engine=user_input["search_engine"],
-            search_engine_version=user_input["search_engine_version"],
-            ident_fdr_psm=user_input["ident_fdr_psm"],
-            ident_fdr_peptide=user_input["ident_fdr_peptide"],
-            ident_fdr_protein=user_input["ident_fdr_protein"],
-            enable_match_between_runs=user_input["enable_match_between_runs"],
-            precursor_mass_tolerance=user_input["precursor_mass_tolerance"],
-            fragment_mass_tolerance=user_input["fragment_mass_tolerance"],
-            enzyme=user_input["enzyme"],
-            allowed_miscleavages=user_input["allowed_miscleavages"],
-            min_peptide_length=user_input["min_peptide_length"],
-            max_peptide_length=user_input["max_peptide_length"],
-            intermediate_hash=str(hashlib.sha1(intermediate.to_string().encode("utf-8")).hexdigest()),
-            comments=user_input["comments_for_submission"],
-        )
-
-        result_datapoint.generate_id()
-        results = dict(ChainMap(*[Module.get_metrics(intermediate, nr_observed) for nr_observed in range(1, 7)]))
-        result_datapoint.results = results
-        result_datapoint.median_abs_epsilon = result_datapoint.results[default_cutoff_min_prec]["median_abs_epsilon"]
-
-        results_series = pd.Series(asdict(result_datapoint))
-
-        return results_series
 
     def add_current_data_point(self, all_datapoints, current_datapoint):
         """Add current data point to all data points and load them from file if empty. TODO: Not clear why is the df transposed here."""
@@ -277,9 +87,12 @@ class Module(ModuleInterface):
         standard_format, replicate_to_raw = parse_settings.convert_to_standard_format(input_df)
 
         # Get quantification data
-        intermediate_data_structure = self.generate_intermediate(standard_format, replicate_to_raw, parse_settings)
+        quant_score = QuantScores(
+            self.precursor_name, parse_settings.species_expected_ratio(), parse_settings.species_dict()
+        )
+        intermediate_data_structure = quant_score.generate_intermediate(standard_format, replicate_to_raw)
 
-        current_datapoint = self.generate_datapoint(
+        current_datapoint = Datapoint.generate_datapoint(
             intermediate_data_structure, input_format, user_input, default_cutoff_min_prec=default_cutoff_min_prec
         )
